more elegant handling of not a directory #18

+++ faces.py
@@ -15,44 +15,46 @@ def check_folder(folder):
         except Exception as e:
             continue
     for actor in folderpaths:
-        try:
-            for path in os.listdir(actor):
-                path_parsed = f"{actor}/{path}"
-                try:
-                    img = cv2.imread(path_parsed)
-                    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-                    faces = face_cascade.detectMultiScale(gray, 1.1, 4)
-                    if len(faces) == 0:
-                        print(f"DELETED: no face found in {path_parsed}")
-                        os.remove(path_parsed)
-                    if len(faces):
-                        # if not os.path.exists(f"{folder}/cropped/"):
-                        #     os.makedirs(f"{folder}/cropped/")
-                        (height, width) = img.shape[:2]
-                        saved_faces = 0
-                        for (x, y, w, h) in faces:
-                            # cv2.rectangle(img, (x, y), (x + w, y + h), (255, 0, 0), 2)
-                            crop_img = img[y : (y + h), x : (x + w)]
-                            if not os.path.exists(
-                                f"./export_preprocessing/cropped/{actor[10:]}/"
-                            ):
-                                os.makedirs(
+        if os.path.isdir(actor):
+            try:
+                for path in os.listdir(actor):
+                    path_parsed = f"{actor}/{path}"
+                    try:
+                        img = cv2.imread(path_parsed)
+                        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+                        faces = face_cascade.detectMultiScale(gray, 1.1, 4)
+                        if len(faces) == 0:
+                            print(f"DELETED: no face found in {path_parsed}")
+                            os.remove(path_parsed)
+                        if len(faces):
+                            # if not os.path.exists(f"{folder}/cropped/"):
+                            #     os.makedirs(f"{folder}/cropped/")
+                            (height, width) = img.shape[:2]
+                            saved_faces = 0
+                            for (x, y, w, h) in faces:
+                                # cv2.rectangle(img, (x, y), (x + w, y + h), (255, 0, 0), 2)
+                                crop_img = img[y : (y + h), x : (x + w)]
+                                if not os.path.exists(
                                     f"./export_preprocessing/cropped/{actor[10:]}/"
+                                ):
+                                    os.makedirs(
+                                        f"./export_preprocessing/cropped/{actor[10:]}/"
+                                    )
+                                cv2.imwrite(
+                                    f"./export_preprocessing/cropped/{actor[10:]}/{path[:-4]}{saved_faces}.jpg",
+                                    crop_img,
                                 )
-                            cv2.imwrite(
-                                f"./export_preprocessing/cropped/{actor[10:]}/{path[:-4]}{saved_faces}.jpg",
-                                crop_img,
-                            )
-                            saved_faces += 1
-                            print(
-                                f"SUCCES: saved cropped version of {actor[10:]}/{path}"
-                            )
-                except Exception as e:
-                    print(f"Could not load {path_parsed} - {e}")
+                                saved_faces += 1
+                                print(
+                                    f"SUCCES: saved cropped version of {actor[10:]}/{path}"
+                                )
+                    except Exception as e:
+                        print(f"Could not load {path_parsed} - {e}")
+                        continue
+            except Exception as e:
+                if e.errno != 20:
+                    print(f"{actor} not a folder: {e}")
                     continue
-        except Exception as e:
-            print(f"Could not load a file - {e}")
-            continue
 
 
 def preprocess_image(folder, width, height, grayscale):
